requirements.txt
anthropic
beautifulsoup4
chromadb
faiss-cpu
feedparser
google-generativeai
gradio
huggingface_hub
ibm-watsonx-ai
ipykernel
ipywidgets
langchain
langchain-core
langchain_ibm
langchain-text-splitters
langchain-openai
langchain-chroma
langchain-community
modal
ollama
openai
plotly
protobuf==3.20.2
psutil
pydub
python-dotenv
requests
ruff
torch==2.2.2
sentence_transformers
speedtest-cli
tavily
wandb
youtube-transcript-api
